feat: Animate show/hide with a duration option
The primary impetus for this change is to allow for animating the show
and hide operations for the terminal window if specified.

This required moving away from using application focus for window
hiding, instead moving towards pushing the window off-screen.

It also required addressing an issue with the way unfocus events were
working when there was a traditional GUI instance of Kitty as well.